Factor out common functionality.
